Add x86 runs without the final dequantization
So the input comes as floats, yet the output stays as bytes as opposed
to going back to float.

Note that due to how memory format works, there is still a transpose
that happens at the end that moves data from NHWC back to NCHW.
So the only difference really is just about no more floating point
magic!