Upgrade to PyTorch 2.4.0
Mainly because the new export mechanism has more-or-less stabilized.

This actually resolves the discrepancy between nn.MaxPool2d and
functional.MaxPool2d when it comes to decompositions: the functional one
used to stubbornly lower to max_pool2d_with_index...

It also prevents two different decompositions for the linear layer
depending on if bias is enabled or not. No more checking for addmm and
mm. Just check for aten.linear and we're set.